# swagger.yaml
openapi: 3.0.0
info:
  title: Admin API
  description: API for managing admin users and chat functionality
  version: 1.0.0
servers:
- url: http://localhost:5764
  description: Local development server
paths:
...
  /assign-course:
    post:
      summary: "Assign course to student"
      description: "Assign a course to a student using their studentId and courseId."
      operationId: "handleAssignCourses"
      tags:
      - "Admin"
      requestBody:
        description: "Data for assigning course to student"
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                courseId:
                  type: string
              required:
              - studentId
              - courseId
      responses:
        '200':
          description: "Course assigned to student successfully"
        '400':
          description: "Student or Course not found"
        '401':
          description: "Unauthorized"
        '500':
          description: "Error assigning course"
  /assign-lesson:
    post:
      summary: "Assign lesson to course"
      description: "Assign a lesson to a course using lessonId and courseId."
      operationId: "handleAssignLessonToCourses"
      tags:
      - "Admin"
      requestBody:
        description: "Data for assigning lesson to course"
        content:
          application/json:
            schema:
              type: object
              properties:
                lessonId:
                  type: string
                courseId:
                  type: string
              required:
              - lessonId
              - courseId
      responses:
        '200':
          description: "Lesson Assigned To Course successfully"
        '400':
          description: "Student or Course not found"
        '401':
          description: "Unauthorized"
        '500':
          description: "Error assigning lesson"
  /assign-quiz:
    post:
      summary: "Assign quiz to lesson"
      description: "Assign a quiz to a lesson using lessonId and quizId."
      operationId: "handleAssignQuizToLesson"
      tags:
      - "Admin"
      requestBody:
        description: "Data for assigning quiz to lesson"
        content:
          application/json:
            schema:
              type: object
              properties:
                lessonId:
                  type: string
                quizId:
                  type: string
              required:
              - lessonId
              - quizId
      responses:
        '200':
          description: "Quiz assigned to lesson successfully"
        '400':
          description: "Quiz or Lesson not found"
        '401':
          description: "Unauthorized"
        '500':
          description: "Error assigning course"
  /students/reports:
    get:
      summary: "Generate reports for all students"
      description: "Generate a report for all students."
      operationId: "handleGetReportsStudents"
      tags:
      - "Admin"
      security:
      - bearerAuth: []
      responses:
        '200':
          description: "Report generated successfully"
        '400':
          description: "Only Admin Can See All Students Detail"
        '500':
          description: "Error generating reports"
  /courses:
    post:
      summary: "Create a new course"
      description: "Create a course using name and description."
      operationId: "createCourse"
      tags:
      - "Course"
      requestBody:
        description: "Course information for creation"
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required:
              - name
              - description
      responses:
        '200':
          description: "Course created successfully"
        '400':
          description: "Only admin can create courses"
        '401':
          description: "Unauthorized"
        '500':
          description: "Error creating course"
  /getCourseProgress:
    post:
      summary: "Get course progress"
      description: "Get the progress of a course for a student using studentId and courseId."
      operationId: "getCourseProgress"
      tags:
      - "Course"
      requestBody:
        description: "Data for retrieving course progress"
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                courseId:
                  type: string
              required:
              - studentId
              - courseId
      responses:
        '200':
          description: "Courses fetched successfully"
        '400':
          description: "Only Admin Can See All Courses Detail"
        '401':
          description: "Unauthorized"
        '500':
          description: "Error fetching courses"
  /courses/{id}:
    get:
      summary: "Get course by ID"
      description: "Get details of a course by its ID."
      operationId: "getCourseById"
      tags:
      - "Course"
      parameters:
      - in: path
        name: id
        required: true
        description: "ID of the course"
        schema:
          type: string
      responses:
        '200':
          description: "Course details retrieved successfully"
        '400':
          description: "Bad request"
        '404':
          description: "Course not found"
        '500':
          description: "Error fetching course"

    put:
      summary: "Update course by ID"
      description: "Update a course's details using its ID and new data."
      operationId: "updateCourse"
      tags:
      - "Course"
      parameters:
      - in: path
        name: id
        required: true
        description: "ID of the course"
        schema:
          type: string
      requestBody:
        description: "Data for updating the course"
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
              required:
              - description
      responses:
        '200':
          description: "Course updated successfully"
        '400':
          description: "Bad request"
        '401':
          description: "Unauthorized"
        '404':
          description: "Course not found"
        '500':
          description: "Error updating course"

    delete:
      summary: "Delete course by ID"
      description: "Delete a course using its ID."
      operationId: "deleteCourse"
      tags:
      - "Course"
      parameters:
      - in: path
        name: id
        required: true
        description: "ID of the course"
        schema:
          type: string
      responses:
        '200':
          description: "Course deleted successfully and removed from students"
        '400':
          description: "Only Admin Can Delete Courses"
        '401':
          description: "Unauthorized"
        '404':
          description: "Course not found"
        '500':
          description: "Error deleting course"

    /quizzes:
    post:
      summary: "Create a new quiz"
      description: "Create a quiz with title, questions, duration, and pass threshold."
      operationId: "handleCreateQuiz"
      tags:
      - "Quiz"
      requestBody:
        description: "Quiz data"
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                questions:
                  type: array
                  items:
                    type: string
                duration:
                  type: integer
                passThreshold:
                  type: integer
              required:
              - title
              - questions
              - duration
              - passThreshold
      responses:
        '200':
          description: "Quiz created successfully"
        '400':
          description: "Bad request"
        '401':
          description: "Unauthorized"
        '500':
          description: "Internal server error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...